Added system_state property to Mfi. The system can now be put into various states such as running, stopped, admin_only and migration. Updated the application controller to reflect this. Added RepaymentStyle to Upload#load_csv and sanitised the order in which things are done. Cleaned up Mfi model and removed irritating validations

+++ lib/tasks/excel.rake
@@ -11,7 +11,12 @@ namespace :excel do
     excel.sheets.each{|sheet|
       puts sheet
       excel.default_sheet=sheet
-      excel.to_csv(File.join("uploads", args.directory, sheet))
+      if File.exists?(File.join("uploads",args.directory, sheet))
+        puts "extracting sheet #{sheet}"
+        excel.to_csv(File.join("uploads", args.directory, sheet))
+      else
+        puts "skipping #{sheet}"
+      end
     }
   end
 end
